fix: correction erreurs compilation Rust - Correction des casts i32 vers i64 dans provider_analytics_service - Gestion Option<bool> dans product_enrichment_service - Ajout annotations de type explicites pour query_scalar - Correction Option<Option<T>> avec flatten() - Correction accès champs tuples et structs - Ajout imports manquants (log::warn, DateTime, Utc, AppError) - Correction problèmes de borrow et move - Correction indexation avec usize au lieu de i32

+++ backend/src/services/dynamic_preparation_time_service.rs
@@ -162,7 +162,7 @@ impl DynamicPreparationTimeService {
 
         times.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
 
-        let count = times.len() as i32;
+        let count = times.len();
         let sum: f64 = times.iter().sum();
         let avg = sum / count as f64;
 
@@ -175,7 +175,7 @@ impl DynamicPreparationTimeService {
         };
 
         // Mettre à jour les stats
-        self.upsert_category_stats(category, avg, median, count).await?;
+        self.upsert_category_stats(category, avg, median, count as i32).await?;
 
         Ok(())
     }
